Rename *MockTest* classes, and combine four test files into two (removing a few tests that were redundant between the files).
